made them byte sized, cause <256 and a constant

+++ src/main/java/teamport/aether/lookup/SlotConstants.java
@@ -3,15 +3,15 @@ package teamport.aether.lookup;
 public final class SlotConstants {
 
     // Slot specific constants, to prevent
-    public static final int ARMOR_START_SLOT = 41;
-    public static final int GLOVES_SLOT = 45;
-    public static final int CAPE_SLOT = 46;
-    public static final int WILDCARD_1_SLOT = 47;
-    public static final int WILDCARD_2_SLOT = 48;
+    public static final byte ARMOR_START_SLOT = 41;
+    public static final byte GLOVES_SLOT = 45;
+    public static final byte CAPE_SLOT = 46;
+    public static final byte WILDCARD_1_SLOT = 47;
+    public static final byte WILDCARD_2_SLOT = 48;
 
     // index specific constants
-    public static final int GLOVES_INDEX= 5;
-    public static final int CAPE_INDEX = 6;
-    public static final int WILDCARD_1_INDEX = 7;
-    public static final int WILDCARD_2_INDEX= 8;
+    public static final byte GLOVES_INDEX= 5;
+    public static final byte CAPE_INDEX = 6;
+    public static final byte WILDCARD_1_INDEX = 7;
+    public static final byte WILDCARD_2_INDEX= 8;
 }
